% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r.R
\name{bubble_sort}
\alias{bubble_sort}
\title{Bubble Sort}
\usage{
bubble_sort(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
sorted numeric vector.
}
\description{
Sorts a numeric vector using the Bubble Sort algorithm
}
