% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r.R
\name{count_sort}
\alias{count_sort}
\title{Counting Sort (Unstable)}
\usage{
count_sort(x)
}
\arguments{
\item{x}{integer vector to be sorted.}
}
\value{
sorted integer vector.
}
\description{
Sorts integer vector using the Counting Sort algorithm.
}
