% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r.R
\name{insertion_sort}
\alias{insertion_sort}
\title{Insertion Sort}
\usage{
insertion_sort(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
sorted numeric vector.
}
\description{
Sorts a numeric vector using the Insertion Sort algorithm
}
