% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r.R
\name{merge_sort}
\alias{merge_sort}
\title{Merge Sort}
\usage{
merge_sort(x)
}
\arguments{
\item{x}{A numeric vector to be sorted.}
}
\value{
A sorted numeric vector.
}
\description{
Sorts numeric vector using the Mergesort algorithm
}
