% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r.R
\name{quick_sort}
\alias{quick_sort}
\title{Quick Sort}
\usage{
quick_sort(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
sorted numeric vector
}
\description{
Sorts a numeric vector using the Quick Sort algorithm
}
