% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r.R
\name{selection_sort}
\alias{selection_sort}
\title{Selection Sort}
\usage{
selection_sort(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
sorted vector
}
\description{
Sorts a numeric vector using the Selection Sort algorithm
}
