% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r.R
\name{selection_sort_CNA}
\alias{selection_sort_CNA}
\title{Selection Sort (but C removes NA (better))}
\usage{
selection_sort_CNA(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
sorted numeric vector
}
\description{
Sorts a numeric vector using the Selection Sort algorithm.
NA and NaN values are removed in C for lower memory usage
}
