% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r.R
\name{shell_sort}
\alias{shell_sort}
\title{Shell Sort}
\usage{
shell_sort(x)
}
\arguments{
\item{x}{A numeric vector to be sorted.}
}
\value{
A sorted numeric vector.
}
\description{
Sorts a numeric vector using the Shell Sort algorithm.
A highly optimized variation of Insertion Sort.
}
